requirements.txt
asgiref==3.8.1
coverage==7.6.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.6
sqlparse==0.5.1
